requirements.txt
requests==2.31.0
lxml==4.9.3
python-dotenv==1.0.0
